+++ src/main/scala/enju/ccg/ml/LogLinearClassifier.scala
@@ -4,23 +4,23 @@ package enju.ccg.ml
   * (implies loss function of log loss)
   */
 trait LogLinearClassifier[L] extends LinearClassifier[L] {
-  def weights: NumericBuffer[Double]
+  def weights: NumericBuffer[Float]
 
-  def labelProbs(examples: Seq[Example[L]]): Array[Double] = {
-    val unnormalized: Array[Double] = examples.map { e =>
-      val p = Math.exp(featureScore(e.featVec))
-      if (p < 1e-100) 1e-100 else p
+  def labelProbs(examples: Seq[Example[L]]): Array[Float] = {
+    val unnormalized: Array[Float] = examples.map { e =>
+      val p = Math.exp(featureScore(e.featVec)).toFloat
+      if (p < 1e-100) 1e-100F else p
     }.toArray
     val z = unnormalized.sum
     unnormalized.map(_ / z)
   }
 }
 
-class ALogLinearClassifier[L](override val weights: NumericBuffer[Double]) extends LogLinearClassifier[L]
+class ALogLinearClassifier[L](override val weights: NumericBuffer[Float]) extends LogLinearClassifier[L]
 
-// class LogLinearClassifier[L](weights: NumericBuffer[Double]) extends LinearClassifier[L](weights) {
-//   def labelProbs(examples: Seq[Example[L]]): Array[Double] = {
-//     val unnormalized: Array[Double] = examples.map { e =>
+// class LogLinearClassifier[L](weights: NumericBuffer[Float]) extends LinearClassifier[L](weights) {
+//   def labelProbs(examples: Seq[Example[L]]): Array[Float] = {
+//     val unnormalized: Array[Float] = examples.map { e =>
 //       val p = Math.exp(featureScore(e.featVec))
 //       if (p < 1e-10) 1e-10 else p
 //     }.toArray
